Add at command handle

+++ Components/uart/serial1.h
@@ -55,17 +55,17 @@ UART_Status_t uart_instance1_deinit(void);
 
 int putchar_instance1(uint8_t c);
 
-uint32_t uart_instance1_can_send(void);
+size_t uart_instance1_can_send(void);
 
 void puts_uart_instance1(char *s);
 
-uint32_t write_uart_instance1(uint8_t *s, uint32_t len);
+size_t write_uart_instance1(uint8_t *s, size_t len);
 
 int getchar_instance1(uint8_t *c);
 
-uint32_t read_uart_instance1(uint8_t *s, uint32_t len);
+size_t read_uart_instance1(uint8_t *s, size_t len);
 
-uint32_t uart_instance1_available(void);
+size_t uart_instance1_available(void);
 
 void uart_instance1_flush(void);
