+++ Components/uart/serial2.h
@@ -59,17 +59,17 @@ UART_Status_t uart_instance2_deinit(void);
 
 int putchar_instance2(uint8_t c);
 
-uint32_t uart_instance2_can_send(void);
+size_t uart_instance2_can_send(void);
 
 void puts_uart_instance2(char *s);
 
-uint32_t write_uart_instance2(uint8_t *s, uint32_t len);
+size_t write_uart_instance2(uint8_t *s, size_t len);
 
 int getchar_instance2(uint8_t *c);
 
-uint32_t read_uart_instance2(uint8_t *s, uint32_t len);
+size_t read_uart_instance2(uint8_t *s, size_t len);
 
-uint32_t uart_instance2_available(void);
+size_t uart_instance2_available(void);
 
 void uart_instance2_flush(void);
