+++ Components/uart/serial5.h
@@ -60,17 +60,17 @@ UART_Status_t uart_instance5_deinit(void);
 
 int putchar_instance5(uint8_t c);
 
-uint32_t uart_instance5_can_send(void);
+size_t uart_instance5_can_send(void);
 
 void puts_uart_instance5(char *s);
 
-uint32_t write_uart_instance5(uint8_t *s, uint32_t len);
+size_t write_uart_instance5(uint8_t *s, size_t len);
 
 int getchar_instance5(uint8_t *c);
 
-uint32_t read_uart_instance5(uint8_t *s, uint32_t len);
+size_t read_uart_instance5(uint8_t *s, size_t len);
 
-uint32_t uart_instance5_available(void);
+size_t uart_instance5_available(void);
 
 void uart_instance5_flush(void);
