+++ Components/uart/serial8.h
@@ -59,17 +59,17 @@ UART_Status_t uart_instance8_deinit(void);
 
 int putchar_instance8(uint8_t c);
 
-uint32_t uart_instance8_can_send(void);
+size_t uart_instance8_can_send(void);
 
 void puts_uart_instance8(char *s);
 
-uint32_t write_uart_instance8(uint8_t *s, uint32_t len);
+size_t write_uart_instance8(uint8_t *s, size_t len);
 
 int getchar_instance8(uint8_t *c);
 
-uint32_t read_uart_instance8(uint8_t *s, uint32_t len);
+size_t read_uart_instance8(uint8_t *s, size_t len);
 
-uint32_t uart_instance8_available(void);
+size_t uart_instance8_available(void);
 
 void uart_instance8_flush(void);
